Major Db migration

+++ backend/poskds/src/main/java/com/anyawalker/poskds/repos/MenuRepo.java
@@ -1,10 +1,10 @@
 package com.anyawalker.poskds.repos;
 
-import com.anyawalker.poskds.models.entities.MenuEntity;
+import com.anyawalker.poskds.models.MenuEntity;
 import org.jspecify.annotations.NonNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MenuRepo extends JpaRepository<@NonNull MenuEntity,@NonNull Long> {
+public interface MenuRepo extends JpaRepository<@NonNull MenuEntity,@NonNull Integer> {
 }
